+++ crates/goose/src/providers/base.rs
@@ -6,7 +6,6 @@ pub use goose_providers::conversation::token_usage::{
     DraftStats, ProviderStats, ProviderUsage, Usage,
 };
 use goose_providers::errors::ProviderError;
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 
 /// Default HTTP timeout for all provider API calls.
@@ -19,10 +18,9 @@ use super::inventory::{default_inventory_identity, InventoryIdentityInput};
 use super::retry::RetryConfig;
 use crate::config::base::ConfigValue;
 use crate::config::{Config, ExtensionConfig, GooseMode};
-use crate::conversation::message::{Message, MessageContent};
-use crate::conversation::Conversation;
 use crate::permission::PermissionConfirmation;
-use crate::utils::safe_truncate;
+use goose_providers::conversation::message::{Message, MessageContent};
+use goose_providers::conversation::Conversation;
 use goose_providers::model::ModelConfig;
 use rmcp::model::Tool;
 use utoipa::ToSchema;
@@ -30,72 +28,10 @@ use utoipa::ToSchema;
 use once_cell::sync::Lazy;
 use std::path::PathBuf;
 use std::pin::Pin;
-use std::sync::{LazyLock, Mutex};
+use std::sync::Mutex;
 
 /// A global store for the current model being used, we use this as when a provider returns, it tells us the real model, not an alias
 pub static CURRENT_MODEL: Lazy<Mutex<Option<String>>> = Lazy::new(|| Mutex::new(None));
-
-fn strip_xml_tags(text: &str) -> String {
-    static BLOCK_RE: LazyLock<Regex> = LazyLock::new(|| {
-        Regex::new(r"(?s)<([a-zA-Z][a-zA-Z0-9_]*)[^>]*>.*?</[a-zA-Z][a-zA-Z0-9_]*>").unwrap()
-    });
-    static TAG_RE: LazyLock<Regex> =
-        LazyLock::new(|| Regex::new(r"</?[a-zA-Z][a-zA-Z0-9_]*[^>]*>").unwrap());
-    let pass1 = BLOCK_RE.replace_all(text, "");
-    TAG_RE.replace_all(&pass1, "").into_owned()
-}
-
-fn extract_short_title(text: &str) -> String {
-    let word_count = text.split_whitespace().count();
-    if word_count <= 8 {
-        return text.to_string();
-    }
-
-    {
-        let mut results = Vec::new();
-        let mut quote_char: Option<char> = None;
-        let mut current = String::new();
-        let mut prev_char: Option<char> = None;
-
-        for ch in text.chars() {
-            match quote_char {
-                None => {
-                    if matches!(ch, '"' | '\'' | '`') {
-                        let after_alnum = prev_char.map(|p| p.is_alphanumeric()).unwrap_or(false);
-                        if !after_alnum {
-                            quote_char = Some(ch);
-                            current.clear();
-                        }
-                    }
-                }
-                Some(q) => {
-                    if ch == q {
-                        let trimmed = current.trim().to_string();
-                        let wc = trimmed.split_whitespace().count();
-                        if (2..=8).contains(&wc) {
-                            results.push(trimmed);
-                        }
-                        quote_char = None;
-                        current.clear();
-                    } else {
-                        current.push(ch);
-                    }
-                }
-            }
-            prev_char = Some(ch);
-        }
-
-        if let Some(title) = results.last() {
-            return title.clone();
-        }
-    }
-
-    if let Some(last) = text.lines().rev().find(|l| !l.trim().is_empty()) {
-        return last.trim().to_string();
-    }
-
-    text.to_string()
-}
 
 pub static MSG_COUNT_FOR_SESSION_NAME_GENERATION: usize = 3;
 
@@ -584,7 +520,7 @@ pub trait Provider: Send + Sync {
                 if !canonical_model
                     .modalities
                     .input
-                    .contains(&crate::providers::canonical::Modality::Text)
+                    .contains(&goose_providers::canonical::Modality::Text)
                 {
                     return None;
                 }
@@ -708,61 +644,6 @@ pub trait Provider: Send + Sync {
             .join("\n")
     }
 
-    /// Generate a session name/description based on the conversation history
-    /// Creates a prompt asking for a concise description in 4 words or less.
-    async fn generate_session_name(
-        &self,
-        session_id: &str,
-        messages: &Conversation,
-    ) -> Result<String, ProviderError> {
-        let context = self.get_initial_user_messages(messages);
-        let preprompt_context = self.get_preprompt_context(messages);
-        let system = crate::prompt_template::render_template(
-            "session_name.md",
-            &std::collections::HashMap::<String, String>::new(),
-        )
-        .map_err(|e| ProviderError::ContextLengthExceeded(e.to_string()))?;
-
-        use super::cli_common::{
-            SESSION_NAME_BEGIN_MARKER, SESSION_NAME_END_MARKER, SESSION_NAME_SUFFIX,
-        };
-
-        let preprompt_section = if preprompt_context.is_empty() {
-            String::new()
-        } else {
-            format!(
-                "---BEGIN BACKGROUND CONTEXT (for understanding only, do NOT base the title on this)---\n{}\n---END BACKGROUND CONTEXT---\n\n",
-                preprompt_context
-            )
-        };
-
-        let user_text = format!(
-            "{}{}\n{}\n{}\n\n{}",
-            preprompt_section,
-            SESSION_NAME_BEGIN_MARKER,
-            context.join("\n"),
-            SESSION_NAME_END_MARKER,
-            SESSION_NAME_SUFFIX,
-        );
-        let message = Message::user().with_text(&user_text);
-        let result = self
-            .complete_fast(session_id, &system, &[message], &[])
-            .await?;
-
-        let raw: String = result
-            .0
-            .content
-            .iter()
-            .filter_map(|c| c.as_text())
-            .collect();
-        let description = strip_xml_tags(&raw)
-            .split_whitespace()
-            .collect::<Vec<_>>()
-            .join(" ");
-
-        Ok(safe_truncate(&extract_short_title(&description), 100))
-    }
-
     /// Configure OAuth authentication for this provider
     ///
     /// This method is called when a provider has configuration keys marked with oauth_flow = true.
@@ -872,87 +753,6 @@ mod tests {
     use super::*;
     use std::collections::HashMap;
     use test_case::test_case;
-
-    #[test]
-    fn test_strip_xml_tags() {
-        assert_eq!(strip_xml_tags("<think>reasoning</think>answer"), "answer");
-        assert_eq!(strip_xml_tags("before<t>mid</t>after"), "beforeafter");
-        assert_eq!(strip_xml_tags("<a>x</a><b>y</b>z"), "z");
-        assert_eq!(strip_xml_tags("no tags here"), "no tags here");
-        assert_eq!(strip_xml_tags("a < b > c"), "a < b > c");
-        assert_eq!(strip_xml_tags("<think>über</think>ok"), "ok");
-        assert_eq!(strip_xml_tags("<think>日本語</think>hello"), "hello");
-        assert_eq!(strip_xml_tags(""), "");
-        assert_eq!(strip_xml_tags("<>stuff</>"), "<>stuff</>");
-        // attributes
-        assert_eq!(
-            strip_xml_tags(r#"<think class="deep">reasoning</think>answer"#),
-            "answer"
-        );
-        // self-closing tags
-        assert_eq!(strip_xml_tags("<br/>self closing"), "self closing");
-        // orphan closing tags
-        assert_eq!(strip_xml_tags("orphan </think> tag"), "orphan  tag");
-        // multiline content
-        assert_eq!(
-            strip_xml_tags("<think>\nline1\nline2\n</think>result"),
-            "result"
-        );
-    }
-
-    #[test]
-    fn test_extract_short_title() {
-        assert_eq!(extract_short_title("List files"), "List files");
-        assert_eq!(
-            extract_short_title(
-                r#"blah blah blah blah blah blah blah blah blah "List files in folder""#
-            ),
-            "List files in folder"
-        );
-        assert_eq!(
-            extract_short_title(
-                "blah blah blah blah blah blah blah blah blah `View current files`"
-            ),
-            "View current files"
-        );
-        assert_eq!(
-            extract_short_title(
-                r#"stuff stuff stuff stuff stuff stuff stuff stuff "Abc title" "Zzz title""#
-            ),
-            "Zzz title"
-        );
-        assert_eq!(
-            extract_short_title(
-                "long long long long long long long long long\nList files in folder"
-            ),
-            "List files in folder"
-        );
-        assert_eq!(
-            extract_short_title(
-                r#"lots of words here and there and more and more "single" final line here"#
-            ),
-            "lots of words here and there and more and more \"single\" final line here"
-        );
-        assert_eq!(extract_short_title("Hello world"), "Hello world");
-        assert_eq!(
-            extract_short_title(
-                r#"1. Analyze the request. 2. The user's message says list files. 3. "List current folder files" fits perfectly. Result: List current folder files"#
-            ),
-            "List current folder files"
-        );
-        assert_eq!(
-            extract_short_title(
-                r#"the user's phrasing is about listing files and the user's intent is clear. "List folder files" is best"#
-            ),
-            "List folder files"
-        );
-        assert_eq!(
-            extract_short_title(
-                "lots of reasoning here about what to call it\nList current folder files"
-            ),
-            "List current folder files"
-        );
-    }
 
     #[test]
     fn test_usage_creation() {
